#13 Rename module (comparable → comparables)
