+++ ndtools/comparison/comparables.py
